+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Implemented the ability to view application logs within the settings screen and share them for troubleshooting purposes.
+
 ## [0.2.0] - 2025-03-25
 
 ### Added